+++ core/src/main/java/ru/origami/common/parallel/EnvironmentContext.java
@@ -1,5 +1,7 @@
 package ru.origami.common.parallel;
 
+import ru.origami.common.environment.Environment;
+
 public class EnvironmentContext {
 
     private static final ThreadLocal<TestEnvironment> current = new ThreadLocal<>();
@@ -9,10 +11,22 @@ public class EnvironmentContext {
     }
 
     public static TestEnvironment getCurrent() {
-        return current.get();
+        TestEnvironment env = current.get();
+
+        if (env == null) {
+            env = Environment.getParallelEnvironmentPool().acquire();
+            current.set(env);
+        }
+
+        return env;
     }
 
     public static void clear() {
-        current.remove();
+        TestEnvironment env = current.get();
+
+        if (env != null) {
+            Environment.getParallelEnvironmentPool().release(env);
+            current.remove();
+        }
     }
 }