+++ core/src/main/java/ru/origami/common/parallel/TestEnvironment.java
@@ -2,18 +2,12 @@ package ru.origami.common.parallel;
 
 import lombok.Getter;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class TestEnvironment {
 
     @Getter
     private final int id;
 
     private boolean isBusy = false;
-
-    @Getter
-    private Map<String, Integer> genericContainerInfos = new HashMap<>();
 
     public TestEnvironment(int id) {
         this.id = id;